--- a/sources/engine/Xenko.Graphics/UIBatch.cs
+++ b/sources/engine/Xenko.Graphics/UIBatch.cs
@@ -215,7 +215,7 @@ namespace Xenko.Graphics
         public void DrawRectangle(ref Matrix worldMatrix, ref Vector3 elementSize, ref Color color, int depthBias)
         {
             // Skip items with null size
-            if (elementSize.Length() < MathUtil.ZeroTolerance)
+            if (elementSize.LengthSquared() < MathUtil.ZeroTolerance)
                 return;
 
             // Calculate the information needed to draw.
@@ -277,7 +277,7 @@ namespace Xenko.Graphics
         private void DrawCube(ref Matrix worldMatrix, ref Vector3 elementSize, ref Color color, int depthBias, bool isReverse)
         {
             // Skip items with null size
-            if (elementSize.Length() < MathUtil.ZeroTolerance)
+            if (elementSize.LengthSquared() < MathUtil.ZeroTolerance)
                 return;
 
             // Calculate the information needed to draw.
@@ -331,7 +331,7 @@ namespace Xenko.Graphics
             if (texture == null) throw new ArgumentNullException(nameof(texture));
 
             // Skip items with null size
-            if (elementSize.Length() < MathUtil.ZeroTolerance)
+            if (elementSize.LengthSquared() < MathUtil.ZeroTolerance)
                 return;
 
             // Calculate the information needed to draw.