--- a/sources/engine/Xenko.Rendering/Streaming/StreamingManager.cs
+++ b/sources/engine/Xenko.Rendering/Streaming/StreamingManager.cs
@@ -438,9 +438,9 @@ namespace Xenko.Streaming
                 return;
 
             // Update resources
-            lock (resources)
+            if (Enabled)
             {
-                if (Enabled)
+                lock (resources)
                 {
                     // Perform synchronization
                     FlushSync();
@@ -558,7 +558,7 @@ namespace Xenko.Streaming
 
         private void FlushSync()
         {
-            for (int i = 0; i < activeStreaming.Count; i++)
+            for (int i = activeStreaming.Count - 1; i >= 0; i--)
             {
                 var resource = activeStreaming[i];
                 if (resource.IsTaskActive)
@@ -566,7 +566,6 @@ namespace Xenko.Streaming
 
                 resource.FlushSync();
                 activeStreaming.RemoveAt(i);
-                i--;
             }
         }