@xen2 2020-03-27
Thanks, looks good to me!
BTW, for next benchmarks, can you try to also use .NET Core 3.1 runtime (if it's available as an option)?
I heard there was some big perf improvements in various area, not sure if during 3.0 or 3.1. I was curious to see if such optimizations are still necessary (and if yes which kind).